requirements.txt
# AWS SDK packages
boto3>=1.34.0
botocore>=1.34.0

# Database connectivity
psycopg2-binary>=2.9.9

# OpenSearch client
opensearch-py>=2.4.0

# File handling and utilities
python-magic>=0.4.27
python-magic-bin>=0.4.14; sys_platform == 'win32'

# Additional utilities
python-dotenv>=1.0.0
requests>=2.31.0
